requirements.txt
ultralytics
opencv-python
scikit-learn
numpy
pillow